mos-core: Assorted v0.5 Fixes and Additions
- Introduce the initial infrastructure into the mos-core
  branch for adding new mOS system calls:

  * a new component (libmos) is added to mOS/tools.
  * The component contains the implementation of the
    user-space -> kernel adaptor.  This is mostly
    mechanical passing of arguments via the syscall()
    routine.  Initially, there are no adapters and so
    this file is mostly empty.
  * The adaptors are built and packaged into a shared
    library (libmos.so).
  * Both the shared library and the uapi/linux/mos.h
    header file are installed in /usr/lib and /usr/include
    (respectively) as part of the RPM packaging and install
    process.

- For mOS sysfs entries, copy the permissions from group to other
  rather than from user.  This allows some mOS sysfs entries to have
  different permissions for root.  For example, lwkmem_debug was
  originally intended to be root writeable but not world writeable.

- Fix some compiler warnings.

- Serialize the mOS/tools build to avoid parallel make race
  conditions for the shared object files.


mos-core: Parallel make Error

I occasionally see this error when doing a parallel mOS build:

  HOSTCC  mOS/tools/lwkctl/lwkctl.o
  CC      arch/x86/kernel/acpi/cstate.o
  HOSTCC  mOS/tools/lwkctl/../lib/mos_cpuset.o
  CC      drivers/acpi/reboot.o
  HOSTCC  mOS/tools/yod/../lib/mos_cpuset.o
  CC      arch/x86/kernel/apic/apic_noop.o
  CC      security/selinux/ss/avtab.o
  ...
fixdep: error opening depfile: mOS/tools/yod/../lib/.mos_cpuset.o.d: No such file or directory
make[4]: *** [mOS/tools/yod/../lib/mos_cpuset.o] Error 2
make[4]: *** Waiting for unfinished jobs....
  CC      arch/x86/kernel/kprobes/opt.o
  CC      arch/x86/kernel/fpu/signal.o

It appears to be a race condition between yod and lwkctl to compile the
mos_cpuset.c file, which was made common when they dynamic CPU
partitioning code was introduced.

Fix this by constraining parallel build for the mOS/tools/*.

This has no measureable affect on the time to build mOS.

+++ mOS/tools/lwkreset/lwkreset.c
@@ -69,7 +69,7 @@ static int get_lwk_processes(char *buff, size_t len)
 
 	if (rc < 0) {
 		LWKRESET_ERR("Could not read \"%s\" (rc = %ld)", MOS_SYSFS_LWKPROCESSES, len);
-	} else if (rc < len) 
+	} else if ((unsigned int)rc < len)
 		buff[rc] = 0; /* force end-of-string */
 
 	fclose(fptr);
@@ -84,7 +84,7 @@ static int get_lwkpid(char **ppidstr)
         unsigned long int pid;
         pid = strtoul(*ppidstr, &endptr, 0);
         *ppidstr = endptr;
-        if (**ppidstr == ',') 
+	if (**ppidstr == ',')
                 *ppidstr += 1;
         return pid;
 }
@@ -97,13 +97,13 @@ void lwkreset_abort(int rc, const char* format, ...)
 	vsprintf(buffer, format, args);
 	fprintf(stderr, "[lwkreset:%d] %s (rc=%d)\n", getpid(), buffer, rc);
 	va_end(args);
-        
+
 	exit(rc);
 }
 
 static void usage(void)
 {
-	int i;
+	unsigned int i;
 
 	printf("Usage: lwkreset [options]\n");
 	printf("Options:\n");
